move shard serialization tests to kernel tests
