+++ tensorflow/python/data/experimental/kernel_tests/serialization/BUILD
@@ -165,23 +165,6 @@ tf_py_test(
 )
 
 tf_py_test(
-    name = "shard_dataset_serialization_test",
-    size = "medium",
-    srcs = ["shard_dataset_serialization_test.py"],
-    tags = [
-        "no_oss",
-        "no_pip",
-        "no_windows",
-    ],
-    deps = [
-        "//tensorflow/python:client_testlib",
-        "//tensorflow/python/data/kernel_tests:checkpoint_test_base",
-        "//tensorflow/python/data/ops:dataset_ops",
-        "@absl_py//absl/testing:parameterized",
-    ],
-)
-
-tf_py_test(
     name = "shuffle_dataset_serialization_test",
     size = "medium",
     srcs = ["shuffle_dataset_serialization_test.py"],
