+++ tensorflow/python/data/kernel_tests/BUILD
@@ -695,6 +695,7 @@ tf_py_test(
     size = "small",
     srcs = ["shard_test.py"],
     deps = [
+        ":checkpoint_test_base",
         ":test_base",
         "//tensorflow/python:client_testlib",
         "//tensorflow/python:errors",
